+++ packages/components/accordion/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @jala-banyu/accordion
 
+## 1.1.0
+
+### Minor Changes
+
+- minor update
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/divider@4.1.0
+  - @jala-banyu/theme@1.4.0
+  - @jala-banyu/use-aria-accordion@1.1.0
+  - @jala-banyu/use-aria-press@1.1.0
+  - @jala-banyu/aria-utils@1.1.0
+  - @jala-banyu/framer-transitions@1.1.0
+  - @jala-banyu/react-utils@1.1.0
+  - @jala-banyu/shared-icons@1.1.0
+  - @jala-banyu/shared-utils@1.1.0
+
 ## 1.0.2
 
 ### Patch Changes